requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
psycopg2-binary>=2.9
django-cors-headers>=4.3
django-filter>=24.2
celery>=5.3
redis>=5.0
qrcode>=7.4
reportlab>=4.0
gunicorn>=21.2
Pillow>=10.0
WeasyPrint>=62.0
django-storages[boto3]>=1.14
boto3>=1.35
twilio>=9.2
